Extract statements_transaction.go
